==> 1_melanoma/3_predict_TCGA_Sweden.Rmd <==
---
title: "Building final models using the MIA features, Survival"
author: "Kevin Wang"
date: "`r paste0('Initiated on 2020 Mar 28, compiled on ', format(Sys.time(), '%Y %b %d'))`"
output:
  html_document:
    code_folding: hide
    fig_height: 7
    fig_width: 10
    toc: yes
    number_sections: true
    toc_depth: 3
    toc_float: yes
editor_options: 
  chunk_output_type: console
---


# Summary

We present two ways of presenting the final model

+ Using the bootstrapped MIA CPOP models as an ensemble of models to make predictions 
+ Using the bootstrapped MIA CPOP models' features as a final features to come up with one singular model for prediction 



```{r}
today = format(Sys.time(), "%Y_%b_%d")
cat("This file was compiled on", today)


suppressPackageStartupMessages({
  library(tidyverse)
  library(CPOP)
  library(glmnet)
  library(yardstick)
  library(survival)
  library(survminer)
  library(patchwork)
  library(fairness)
})

theme_set(theme_classic(12) +
            theme(legend.position = "bottom"))
```


## data

```{r}
file_version = "2020_Apr_25"
load(paste0(
  "../data/processed_data/Melanoma4_binary_",
  file_version, ".RData"
))


load(paste0(
  "../data/processed_data/Melanoma4_survival_",
  file_version, ".RData"
))
```


```{r}
list_lratio_binary = list_raw_data_binary %>%
  purrr::map(t) %>%
  purrr::map(CPOP::pairwise_col_diff)

list_lratio_survival = list_raw_data %>%
  purrr::map(t) %>%
  purrr::map(CPOP::pairwise_col_diff)
```




## Customised functions
```{r}
join_tcga_sample_data = function(dat, by){
  dat %>% 
    left_join(tcga_sample_data, by = c("samples" = "sample_id")) %>% 
    dplyr::filter(naive_tumor_stage == "stageIII", !str_detect(samples, "TCGA-EE")) %>% 
    dplyr::mutate(time = time/365) %>% 
    dplyr::mutate(
      tcga_median_class = case_when(
        status == 0 & time > median(time) ~ "Good",
        status == 1 & time < median(time) ~ "Poor",
        TRUE ~ NA_character_))
}

join_sweden_sample_data = function(dat, by){
  dat %>% 
    left_join(sweden_sample_data, by = c("samples" = "sample_id")) %>% 
    dplyr::mutate(time = dss/365, status = dss_dead) %>% 
    dplyr::mutate(
      sweden_median_class = case_when(
        status == 0 & time > median(time) ~ "Good",
        status == 1 & time < median(time) ~ "Poor",
        TRUE ~ NA_character_))
}

binary_split = function(prob){
  ifelse(prob > 0.5, "Poor", "Good")
}

trinary_split = function(prob){
  case_when(
    prob < 0.3 ~ "Good",
    prob < 0.7 ~ "Moderate",
    prob < 1 ~ "Poor"
  )
}
```


# Survival times distributional comparison 


```{r}
list_surv_times = lst(
  mia = tibble(
    time = rfs_array_sample_data_reduced$RFS, 
    dead = rfs_array_sample_data_reduced$dead), 
  tcga = tibble(
    time = tcga_sample_data$time/365, 
    dead = tcga_sample_data$status),
  sweden = tibble(
    time = sweden_sample_data$dss/365, 
    dead = sweden_sample_data$dss_dead))

surv_times_tbl = list_surv_times %>% bind_rows(.id = "data_source")

list_survival_lassoy = list(
  ns = Surv(rfs_ns_sample_data_reduced$RFS + 1,
            rfs_ns_sample_data_reduced$dead),
  array = Surv(rfs_array_sample_data_reduced$RFS + 1,
               rfs_array_sample_data_reduced$dead),
  tcga = Surv(tcga_sample_data$time,
               tcga_sample_data$status),
  sweden = Surv(sweden_sample_data$dss,
               sweden_sample_data$dss_dead))
```

# Single CPOP model from MIA (saved)
```{r}
# set.seed(12)
# 
# mia_cpop_binary = CPOP::cpop_model(
#   z1 = list_lratio_binary$ns,
#   z2 = list_lratio_binary$array,
#   y1 = list_binary_lassoy$ns,
#   y2 = list_binary_lassoy$array,
#   family = "binomial",
#   alpha = 0.1,
#   n_features = 20)
# 
# save(mia_cpop_binary, file = "../data/processed_data/shiny_models_2021_Nov_02.RData")
```

```{r}
load(file = "../data/processed_data/shiny_models_2021_Nov_02.RData")
```


## TCGA 
```{r}
mia_single_pred_tcga = predict_cpop(mia_cpop_binary, newz = list_lratio_survival$tcga) %>% 
  join_tcga_sample_data() %>% 
  as.data.frame()

table(mia_single_pred_tcga$cpop_model_avg_class)

singlemia_tcga_km = survminer::ggsurvplot(
  fit = survfit(Surv(time, status) ~ cpop_model_avg_class, 
                data = mia_single_pred_tcga), 
  data = mia_single_pred_tcga, 
  break.time.by = 5,
  pval = TRUE, risk.table = "nrisk_cumevents")

singlemia_tcga_km2 = singlemia_tcga_km$plot +
  ggsci::scale_colour_d3()

singlemia_tcga_km3 = singlemia_tcga_km$table +
  ggsci::scale_colour_d3()

coxph(Surv(time, status) ~ cpop_model_avg_prob + age + gender, 
      data = mia_single_pred_tcga) %>% 
  summary()

singlemia_tcga_km_plots = singlemia_tcga_km2 + singlemia_tcga_km3 +
  patchwork::plot_layout(ncol = 1, byrow = FALSE, heights = c(1.5, 1))

singlemia_tcga_km_plots

ggsave(plot = print(singlemia_tcga_km_plots),
       filename = paste0("../figures/singlemia_tcga_km_plots_", today, ".eps"),
       height = 6, width = 8)
```

Fairness calculations is only relevant for TCGA samples with a well defined binary classifications. We will use the definition: 

Good: survived more than the median survival time (26.9 mo, or 2.24 years). 
Poor: survived less than the median survival time.

```{r, fig.height = 4, fig.width = 6}
print(mia_single_pred_tcga$time %>% median)

equal_odds(
  data  = mia_single_pred_tcga %>% 
    dplyr::mutate(
      tcga_median_class = case_when(
        status == 0 & time > median(time) ~ "Good",
        status == 1 & time < median(time) ~ "Poor",
        TRUE ~ NA_character_)) %>% 
    dplyr::filter(complete.cases(tcga_median_class)),
  outcome = 'tcga_median_class',
  probs   = 'cpop_model_avg_prob',
  group   = 'gender',
  cutoff  = 0.5)
```


## Sweden 
```{r}
mia_single_pred_sweden = predict_cpop(mia_cpop_binary, newz = list_lratio_survival$sweden) %>% 
  join_sweden_sample_data() %>% 
  as.data.frame()

table(mia_single_pred_sweden$cpop_model_avg_class)

singlemia_sweden_km = survminer::ggsurvplot(
  fit = survfit(Surv(time, status) ~ cpop_model_avg_class, 
                data = mia_single_pred_sweden), 
  data = mia_single_pred_sweden, 
  break.time.by = 5,
  pval = TRUE, risk.table = "nrisk_cumevents")

singlemia_sweden_km2 = singlemia_sweden_km$plot +
  ggsci::scale_colour_d3()

singlemia_sweden_km3 = singlemia_sweden_km$table +
  ggsci::scale_colour_d3()

coxph(Surv(time, status) ~ cpop_model_avg_prob + age + gender, 
      data = mia_single_pred_sweden)

singlemia_sweden_km_plots = singlemia_sweden_km2 + singlemia_sweden_km3 +
  patchwork::plot_layout(ncol = 1, byrow = FALSE, heights = c(1.5, 1))

singlemia_sweden_km_plots

ggsave(plot = print(singlemia_sweden_km_plots),
       filename = paste0("../figures/singlemia_sweden_km_plots_", today, ".eps"),
       height = 6, width = 8)
```

Fairness calculations is only relevant for Sweden samples with a well defined binary classifications. We will use the definition: 

Good: survived more than the median survival time (17.6 mo, or 1.46 years). 
Poor: survived less than the median survival time.

```{r, fig.height = 4, fig.width = 6}
print(median(mia_single_pred_sweden$time))

equal_odds(
  data  = mia_single_pred_sweden %>% 
    dplyr::mutate(
      sweden_median_class = case_when(
        status == 0 & time > median(time) ~ "Good",
        status == 1 & time < median(time) ~ "Poor",
        TRUE ~ NA_character_)) %>% 
    dplyr::filter(complete.cases(sweden_median_class)),
  outcome = 'sweden_median_class',
  probs   = 'cpop_model_avg_prob',
  group   = 'gender',
  cutoff  = 0.5)
```

# Features evaluations
```{r}
## We stored all features data in this file.
load("../data/processed_data/ns_validation60_2022_Feb_07.RData")
ns_feature_data = ns_feature_data %>% 
  as_tibble() %>% 
  dplyr::mutate(
    accession_short = accession %>% 
      str_extract("NM_\\s*(.*?)\\s*[.]") %>% 
      str_remove("[.]"))
```

```{r}
mia_cpop_binary_genes = mia_cpop_binary$feature %>% 
  str_split("--") %>% unlist %>% unique
```

```{r, eval = FALSE}
library(biomaRt)
gene_list <- c(ns_feature_data$accession_short)
human <- useMart("ensembl", dataset="hsapiens_gene_ensembl")
gene_coords=getBM(
  attributes=c("hgnc_symbol","refseq_mrna", "start_position","end_position"), 
  filters="refseq_mrna", 
  values=gene_list, mart=human)
gene_coords$length = gene_coords$end_position - gene_coords$start_position
saveRDS(gene_coords, file = "../data/processed_data/nanostring_gene_length_2022_Feb_07.RData")
```


```{r}
gene_lengths = readRDS("../data/processed_data/nanostring_gene_length_2022_Feb_07.RData") %>% 
  as_tibble() %>% 
  group_by(hgnc_symbol, refseq_mrna) %>% 
  summarise(length = median(length))

ns_feature_data_length = ns_feature_data %>% 
  inner_join(gene_lengths, 
             by = c("gene_name" = "hgnc_symbol", 
                    "accession_short" = "refseq_mrna")) %>% 
  dplyr::mutate(cpop_selected = gene_name %in% mia_cpop_binary_genes)
```

```{r}
library(ggbeeswarm)

ns_feature_data_length %>% 
  ggplot(aes(x = cpop_selected, y = length)) +
  geom_violin(width = 0.5) +
  geom_boxplot(width = 0.2) +
  geom_quasirandom(width = 0.1) +
  scale_y_log10(label = scales::label_comma()) +
  labs(x = "CPOP selected", 
       y = "Gene length")

wilcox.test(log10(length) ~ cpop_selected, 
            data = ns_feature_data_length)
```

```{r}
gene_median_plotdf = list_raw_data_binary %>% 
  purrr::map_dfr(
    .f = ~ {
      apply(.x, 1, median) %>% 
        tibble(
          gene_name = names(.),
          gene_expr_median = .) %>% 
        dplyr::mutate(
          cpop_selected = gene_name %in% mia_cpop_binary_genes)
    }, .id = "data_name") %>% 
  dplyr::mutate(
    data_name = case_when(
      data_name == "array" ~ "MIA-Microarray",
      data_name == "ns" ~ "MIA-NanoString",
      data_name == "tcga" ~ "TCGA",
      data_name == "sweden" ~ "Sweden")
  )
   

gene_median_plotdf %>% 
  ggplot(aes(x = cpop_selected, y = gene_expr_median)) +
  geom_violin(width = 0.5) +
  geom_boxplot(width = 0.2) +
  geom_quasirandom(size = 0.5, width = 0.1) + 
  facet_wrap(~data_name, scales =  "free_y") + 
  labs(x = "CPOP selected", 
       y = "Median of gene expression")

gene_median_plotdf %>% 
  group_by(data_name) %>% 
  tidyr::nest() %>% 
  dplyr::mutate(
    wilcox_tidy = purrr::map_dfr(
      .x = data, 
      .f = ~ broom::tidy(wilcox.test(gene_expr_median ~ cpop_selected, data = .x)))
    ) %>% 
  dplyr::select(-data)
```

```{r}
ratio_median_plotdf = list_lratio_binary %>% 
  purrr::map_dfr(
    .f = ~ {
      apply(.x, 2, median) %>% 
        tibble(
          ratio_name = names(.),
          ratio_median = .) %>% 
        dplyr::mutate(
          cpop_selected = ratio_name %in% mia_cpop_binary$feature)
    }, .id = "data_name") %>% 
  dplyr::mutate(
    data_name = case_when(
      data_name == "array" ~ "MIA-Microarray",
      data_name == "ns" ~ "MIA-NanoString",
      data_name == "tcga" ~ "TCGA",
      data_name == "sweden" ~ "Sweden")
  )
   

ratio_median_plotdf %>% 
  ggplot(aes(x = cpop_selected, y = ratio_median)) +
  geom_violin(width = 0.5) +
  geom_boxplot(width = 0.2) +
  # geom_quasirandom(size = 0.5, width = 0.1) + 
  facet_wrap(~data_name, scales =  "free_y") + 
  labs(x = "CPOP selected", 
       y = "Median of log-ratio")

ratio_median_plotdf %>% 
  group_by(data_name) %>% 
  tidyr::nest() %>% 
  dplyr::mutate(
    wilcox_tidy = purrr::map_dfr(
      .x = data, 
      .f = ~ broom::tidy(wilcox.test(ratio_median ~ cpop_selected, data = .x)))
    ) %>% 
  dplyr::select(-data)
```
