==> 1_melanoma/4_predict_MIAvalidation60.Rmd <==
---
title: "MIA - March validation 60 samples data cleaning"
author: "Kevin Wang"
date: "`r paste0('Initiated on 2020 Feb 16, compiled on ', format(Sys.time(), '%Y %b %d'))`"
output:
  html_document:
    code_folding: hide
    fig_height: 10
    fig_width: 10
    toc: yes
    number_sections: true
    toc_depth: 3
    toc_float: yes
editor_options: 
  chunk_output_type: console
---


# Summary


# Loading 

## Packages

```{r}
today = format(Sys.time(), "%Y_%b_%d")
cat("This file was compiled on", today)


suppressPackageStartupMessages({
  library(tidyverse)
  library(NanoStringQCPro)
  library(survminer)
  library(survival)
  library(ggsci)
  library(janitor)
})

theme_set(theme_classic(20) +
            theme(legend.position = "bottom"))

boxplot_tbl = function(x, index = 1){
  means = apply(x, index, mean, na.rm = TRUE)
  medians = apply(x, index, median, na.rm = TRUE)
  q1 = apply(x, index, quantile, 0.25, na.rm = TRUE)
  q3 = apply(x, index, quantile, 0.75, na.rm = TRUE)
  
  tbl = tibble(
    object = as_factor(dimnames(x)[[index]]),
    means, medians, q1, q3)
  
  return(tbl)
}

```



## Raw NanoString data
```{r}
rcc_files = list.files("../data/raw_data/validation60_15_March_2020/", 
                       recursive = TRUE, full.names = TRUE,
                       pattern = "[.]RCC")

ns_v60_data = newRccSet(rccFiles = rcc_files,
                        rlf ="../data/raw_data/Pupo_1_C12874.RLF")


cat("If the following dimensions match with the number of RCC files above, then we have read in all files properly.")
dim(ns_v60_data)
stopifnot(ncol(ns_v60_data) == length(rcc_files))
```

## Previous 45 samples  
```{r}
file_version = "2020_Apr_25"
load(paste0(
  "../data/processed_data/Melanoma4_binary_",
  file_version, ".RData"
))

load(paste0(
  "../data/processed_data/Melanoma4_survival_",
  file_version, ".RData"
))
```


# Processing sample, feature, exprs data for NanoString


## Feature data

+ We will extract the gene probes data to ease the processing. 

```{r}
ns_feature_data = fData(ns_v60_data) %>% 
  janitor::clean_names()

ns_all_genes = ns_feature_data$gene_name

ns_endo_genes = ns_feature_data %>% 
  dplyr::filter(code_class == "Endogenous") %>% 
  dplyr::pull(gene_name)

ns_hk_genes = ns_feature_data %>% 
  dplyr::filter(code_class == "Housekeeping") %>% 
  dplyr::pull(gene_name)

# nsHkGenes = c("RERE", "CENPB", "CTBP1", "GNB2L1", "SNRPD2", "UQCR")
```


## Sample data 


```{r}
ns_v60_raw_sample_data = pData(ns_v60_data) %>% 
  janitor::clean_names() %>% 
  dplyr::mutate(date = lubridate::ymd(date),
                cartridge_id = as.character(cartridge_id),
                sample_id = trimws(sample_id)) %>% 
  remove_empty("cols") %>% as_tibble()

ns_v60_raw_sample_data %>% glimpse()

## One of the samples had a mis-label. See May 18 email by GP and KW, https://trello.com/c/sizkc4Ku. 
ns_v60_raw_sample_data = ns_v60_raw_sample_data %>% 
  dplyr::mutate(sample_id = ifelse(sample_id == "50088", "500888", sample_id))

raw_sample_id_mapping_data = readxl::read_excel("../data/raw_data/validation60_15_March_2020/Cohort A NanostringAssay sample Details.xlsx") %>% janitor::clean_names() %>% 
  dplyr::select(pin, gender, vial_barcode, nanostring_sample_id, date_of_nanostring_run_12, other)

raw_clinical_data1 = readxl::read_excel("../data/raw_data/validation60_15_March_2020/cohort a survival_Hazel_21May2020.xlsx", sheet = 1) %>% janitor::clean_names() %>% 
  remove_empty("cols") %>% 
  dplyr::transmute(
    pin,gender, date_tissue_banked, 
    date_last_fup, status_last_fup, 
    stage = stage_at_bank_7th_ed,
    date_first_distant_met_post_banking,
    date_first_recurrence_post_banking,
    date_first_recurrence_post_banking_corrected = case_when(
      complete.cases(date_first_recurrence_post_banking) ~ date_first_recurrence_post_banking, 
      status_last_fup == "Alive NSR" & is.na(date_first_recurrence_post_banking) ~ date_last_fup
    ),
    age = difftime(date_tissue_banked, dob) %>% lubridate::time_length("year"))

raw_clinical_data1 %>% glimpse

## No duplicates after cleaning the clinical data
raw_clinical_data1 %>% janitor::get_dupes(pin)

ns_v60_clean_sample_data = ns_v60_raw_sample_data %>% 
  left_join(raw_sample_id_mapping_data, by = c("sample_id" = "nanostring_sample_id")) %>% 
  left_join(raw_clinical_data1, by = c("pin" = "pin", "gender" = "gender")) %>%
  as_tibble() %>% 
  dplyr::mutate(
    status_last_fup = case_when(
      pin == "41036" ~ "Alive NSR",
      TRUE ~ status_last_fup), ## 2020 June 10, email with GP and Hazel
    RFS = as.integer(difftime(date_first_recurrence_post_banking_corrected,
                              date_tissue_banked,
                              units = "days")/30),
    rfs_class = case_when(
      (RFS >= 4*12 & status_last_fup == "Alive NSR") ~ "Good",
      (RFS <= 1*12 & status_last_fup == "Dead, melanoma") ~ "Poor",
      TRUE ~ NA_character_
    ) %>% as.factor(),
    OS = as.integer(
      difftime(date_last_fup,
               date_tissue_banked,
               units = "days")/30),
    dead = ifelse(str_detect(status_last_fup, "Alive"), 0, 1),
    os_class = case_when(
      (OS >= 4*12 & status_last_fup == "Alive NSR") ~ "Good",
      (OS <= 1*12 & status_last_fup == "Dead, melanoma") ~ "Poor",
      TRUE ~ NA_character_
    ) %>% as.factor()) %>% 
  group_by(sample_id) %>% 
  dplyr::mutate(dupes_v60 = n(), 
                numbering_dupes = row_number(),
                in_mia45 = sample_id %in% colnames(list_raw_data_binary$ns),
                plot_keep = (numbering_dupes == dupes_v60) & !in_mia45) %>% 
  ungroup()

ns_v60_clean_sample_data %>% glimpse()
```


# QC check (Combined with previous 45 samples)
```{r}
ns_v60_exprs_data = log2(exprs(ns_v60_data) + 1)
rownames(ns_v60_exprs_data) = ns_feature_data$gene_name %>% 
  toupper %>%
  str_replace_all("-", "_") %>%
  str_replace_all("[.]", "_")

ns_v60_exprs_data = ns_v60_exprs_data[ns_v60_exprs_data %>% rownames %>% sort,]
```

# Save data
```{r}
save(
  ns_feature_data,
  ns_v60_clean_sample_data,
  ns_v60_exprs_data,
  file = "../data/processed_data/ns_validation60_2022_Feb_07.RData")
```

# Running CPOP model 
```{r}
list_raw_data_binary = list_raw_data_binary %>% purrr::map(t)

list_lratio_binary = purrr::map(list_raw_data_binary, CPOP::pairwise_col_diff)

# c("CPN1", "CXCL13", "CXCR6", "CYP2J2", "CYorf15A", "DARC", "DHRS9", "DNAH2", "CXCL10", "CXCL9", "CYBB", "DCT")
```


```{r}
load("../data/processed_data/shiny_models_2021_Nov_02.RData")

ns_v60_lratio = ns_v60_exprs_data %>% t %>% CPOP::pairwise_col_diff()
```


# Introduce missing values on the genes 

```{r}
faulty_genes = c("CPN1", "CXCL13", "CXCR6", "CYP2J2", "CYorf15A", "DARC", "DHRS9", "DNAH2", "CXCL10", "CXCL9", "CYBB", "DCT") %>% toupper()

cpop_lr2genes = stringr::str_split(mia_cpop_binary$feature, "--") %>% unlist %>% unique %>% sort

gplots::venn(
  lst(faulty_genes, 
      cpop_lr2genes))

ns_v60_exprs_data_miss = t(ns_v60_exprs_data)[,colnames(list_raw_data_binary$array)]
ns_v60_exprs_data_miss[,intersect(faulty_genes, colnames(ns_v60_exprs_data_miss))] = NA

ns_v60_exprs_data_impute = CPOP::impute_cpop(cpop_result = mia_cpop_binary, 
                  x1 = list_raw_data_binary$array, 
                  x2 = list_raw_data_binary$ns, 
                  newx = ns_v60_exprs_data_miss)

ns_v60_lratio = ns_v60_exprs_data_impute %>% CPOP::pairwise_col_diff()
```



```{r}
comb_pred_tbl = mia_cpop_binary %>% 
  CPOP::predict_cpop(newz = ns_v60_lratio) %>%
  left_join(ns_v60_clean_sample_data, by = c("samples" = "file_name")) %>% 
  dplyr::mutate(
    cpop_pred = cpop_model_avg,
    cpop_prob = CPOP::expit(cpop_model_avg),
    cpop_pred_class = ifelse(cpop_model_avg > 0, "Poor", "Good"),
    dead = coalesce(dead, 0))

table(comb_pred_tbl$cpop_pred_class)

comb_pred_tbl %>% 
  ggplot(aes(x = CPOP::expit(cpop_model1), 
             y = CPOP::expit(cpop_model2),
             colour = rfs_class)) +
  geom_point(size = 3) +
  xlim(0, 1) +
  ylim(0, 1) +
  geom_abline(slope = 1, intercept = 0, colour = "red") +
  labs(x = "Prediction from MIA - 2015, Array",
       y = "Prediction from MIA - 2018, NS")

comb_pred_tbl$cpop_pred_class %>% table

table(comb_pred_tbl$cpop_pred_class, 
      comb_pred_tbl$os_class)

fit = survfit(Surv(RFS/12, dead) ~ cpop_pred_class, 
              data = comb_pred_tbl %>% 
                dplyr::filter(plot_keep))

fit

singlemia_validate60_km = ggsurvplot(fit, pval = TRUE, 
                                     break.time.by = 2,
                                     risk.table = "nrisk_cumevents")

singlemia_validate60_km2 = singlemia_validate60_km$plot
singlemia_validate60_km3 = singlemia_validate60_km$table

library(patchwork)

singlemia_validate60_km_plots = singlemia_validate60_km2 + singlemia_validate60_km3 +
  patchwork::plot_layout(ncol = 1, byrow = FALSE, heights = c(1.5, 1))

ggsave(print(singlemia_validate60_km_plots), 
       filename = paste0("../figures/validation_60_surv_pred_", today, ".eps"),
       width = 8, height = 6)

coxph(Surv(RFS, dead) ~ cpop_pred + age, data = comb_pred_tbl)
```

# Session Info 
```{r}
sessioninfo::session_info()
```
