==> 1_melanoma/4_2_visualise_all_five_melanoma_data.Rmd <==
---
title: "Visualise 5 melanoma data"
author: "Kevin Wang"
date: "`r paste0('Initiated on 2022 Apr 03, compiled on ', format(Sys.time(), '%Y %b %d'))`"
output:
  html_document:
    code_folding: hide
    fig_height: 10
    fig_width: 10
    toc: yes
    number_sections: true
    toc_depth: 3
    toc_float: yes
editor_options: 
  chunk_output_type: console
---

# Summary 

This document will visualise the distribution of all five melanoma data. 

# Loading 

## Packages

```{r}
today = format(Sys.time(), "%Y_%b_%d")
cat("This file was compiled on", today)


suppressPackageStartupMessages({
  library(tidyverse)
  library(patchwork)
})

theme_set(theme_classic(12) +
            theme(legend.position = "bottom"))
```

## Data

```{r}
file_version = "2020_Apr_25"
load(paste0(
  "../data/processed_data//Melanoma4_survival_",
  file_version, ".RData"
))
load("../data/processed_data/ns_validation60_2022_Feb_07.RData")
```

```{r}
list_five_data = c(list_raw_data,
                 list(ns_v60 = ns_v60_exprs_data[rownames(list_raw_data$ns),])) %>% 
  purrr::map(t)

list_five_lratio = purrr::map(list_five_data, CPOP::pairwise_col_diff)
```


```{r}
boxplot_tbl = function(x, index = 1){
  means = apply(x, index, mean, na.rm = TRUE)
  medians = apply(x, index, median, na.rm = TRUE)
  q1 = apply(x, index, quantile, 0.25, na.rm = TRUE)
  q3 = apply(x, index, quantile, 0.75, na.rm = TRUE)
  
  tbl = tibble(
    object = as_factor(dimnames(x)[[index]]),
    means, medians, q1, q3)
  
  return(tbl)
}

list_ge_boxplot_tbl = purrr::map_dfr(
  list_five_data, boxplot_tbl, 
  index = 1, .id = "data_source") %>% 
  dplyr::mutate(
    data_source = fct_recode(data_source, 
                             "MIA - Microarray" = "array",
                             "MIA - NanoString" = "ns",
                             "TCGA" = "tcga",
                             "Sweden" = "sweden",
                              "MIA - Validation" = "ns_v60"),
    samples = paste0(data_source, "_", object))


ge_boxplot = list_ge_boxplot_tbl %>%
  ggplot(aes(x = samples, y = medians,
             colour = data_source)) +
  geom_point(size = 2) +
  geom_linerange(aes(ymin = q1,
                     ymax = q3), size = 0.5) +
  scale_colour_brewer(palette = "Set1") +
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank(), legend.position = "bottom") +
  labs(title = "", x = "", y = "")
#######################################
list_lr_boxplot_tbl = purrr::map_dfr(
  list_five_lratio, boxplot_tbl, 
  index = 1, .id = "data_source") %>% 
  dplyr::mutate(
    data_source = fct_recode(data_source, 
                             "MIA - Microarray" = "array",
                             "MIA - NanoString" = "ns",
                             "TCGA" = "tcga",
                             "Sweden" = "sweden",
                             "MIA - Validation" = "ns_v60"),
    samples = paste0(data_source, "_", object))


lr_boxplot = list_lr_boxplot_tbl %>%
  ggplot(aes(x = samples, y = medians,
             colour = data_source)) +
  geom_point(size = 2) +
  geom_linerange(aes(ymin = q1,
                     ymax = q3), size = 0.5) +
  scale_colour_brewer(palette = "Set1") +
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank(), legend.position = "none") +
  labs(title = "", x = "", y = "",
       colour = "")

ge_boxplot + lr_boxplot + patchwork::plot_layout(nrow = 2)

ggsave(filename = paste0("../figures/ge_lr_boxplot_", today, ".eps"),
       plot = ge_boxplot + lr_boxplot + patchwork::plot_layout(nrow = 2), 
       width = 12, height = 8)
```
